+++ core/docs/changelog.md
@@ -1,4 +1,10 @@
 # TBD
+### Changes
+* The API container launcher now has two methods, `LaunchWithCustomVersion` and `LaunchWithDefaultVersion`
+
+### Breaking Changes
+* The API container launcher's `Launch` method has been renamed
+    * Users should use either `LaunchWithCustomVersion` or `LaunchWithDefaultVersion` depending on their needs
 
 # 1.30.0
 ### Changes